modified getFileLines() and performSort() methods
now getFileLines() gives all the lines in a object if only there is no
error while reading file. if there is any error reading file then it
will pass the error message in an object to the callback

perform sort does not need exitsSync anymore. it uses readFile
function's error to determine the error